Add sylius theme and form layout styles
This commit introduces sylius-related theme and form layout styles. The update includes several SCSS files for themes and elements such as buttons and footers, a number of new form layout templates for addresses, logins, and country codes, among others. This update improves the appearance and functionality of the forms and adds a consistent aesthetic to the project.